+++ packages/sst/src/constructs/AstroSite.ts
@@ -22,7 +22,6 @@ const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
  * ```
  */
 export class AstroSite extends SsrSite {
-
   protected initBuildConfig() {
     return {
       serverBuildOutputFile: "dist/server/entry.mjs",
@@ -36,24 +35,30 @@ export class AstroSite extends SsrSite {
     const serverDir = path.join(this.props.path, "dist/server");
     const clientDir = path.join(this.props.path, "dist/client");
     if (!fs.existsSync(serverDir) || !fs.existsSync(clientDir)) {
-      throw new Error(`Build output inside "dist/" does not contain the "server" and "client" folders. Make sure Server-side Rendering (SSR) is enabled in your Astro app. If you are looking to deploy the Astro app as a static site, please use the StaticSite construct — https://docs.sst.dev/constructs/StaticSite`);
+      throw new Error(
+        `Build output inside "dist/" does not contain the "server" and "client" folders. Make sure Server-side Rendering (SSR) is enabled in your Astro app. If you are looking to deploy the Astro app as a static site, please use the StaticSite construct — https://docs.sst.dev/constructs/StaticSite`
+      );
     }
 
     super.validateBuildOutput();
   }
 
   protected createFunctionForRegional(): lambda.Function {
-    const { defaults, environment } = this.props;
+    const { defaults, environment, bind } = this.props;
 
     // Bundle code
     const handler = this.isPlaceholder
-      ? path.resolve(__dirname, "../support/ssr-site-function-stub/index.handler")
+      ? path.resolve(
+          __dirname,
+          "../support/ssr-site-function-stub/index.handler"
+        )
       : path.join(this.props.path, "dist", "server", "entry.handler");
 
     // Create function
     const fn = new Function(this, `ServerFunction`, {
       description: "Server handler",
       handler,
+      bind,
       logRetention: "three_days",
       runtime: "nodejs16.x",
       memorySize: defaults?.function?.memorySize || "512 MB",
@@ -78,8 +83,7 @@ export class AstroSite extends SsrSite {
     if (this.isPlaceholder) {
       bundlePath = path.resolve(__dirname, "../support/ssr-site-function-stub");
       handler = "index.handler";
-    }
-    else {
+    } else {
       // Create a directory that we will use to create the bundled version
       // of the "core server build" along with our custom Lamba server handler.
       const outputPath = path.resolve(
@@ -90,7 +94,9 @@ export class AstroSite extends SsrSite {
       );
 
       const result = esbuild.buildSync({
-        entryPoints: [path.join(this.props.path, this.buildConfig.serverBuildOutputFile)],
+        entryPoints: [
+          path.join(this.props.path, this.buildConfig.serverBuildOutputFile),
+        ],
         target: "esnext",
         format: "esm",
         platform: "node",
@@ -109,11 +115,16 @@ export class AstroSite extends SsrSite {
 
       if (result.errors.length > 0) {
         result.errors.forEach((error) => console.error(error));
-        throw new Error(`There was a problem bundling the function code for the ${this.id} AstroSite.`);
+        throw new Error(
+          `There was a problem bundling the function code for the ${this.id} AstroSite.`
+        );
       }
 
       // Create package.json
-      fs.writeFileSync(path.join(outputPath, "package.json"), `{"type":"module"}`);
+      fs.writeFileSync(
+        path.join(outputPath, "package.json"),
+        `{"type":"module"}`
+      );
 
       bundlePath = outputPath;
       handler = "entry.handler";