+++ packages/sst/src/constructs/SsrSite.ts
@@ -55,9 +55,21 @@ export type SsrBuildConfig = {
   siteStub: string;
 };
 
-export interface SsrDomainProps extends BaseSiteDomainProps { }
-export interface SsrCdkDistributionProps extends BaseSiteCdkDistributionProps { }
+export interface SsrDomainProps extends BaseSiteDomainProps {}
+export interface SsrCdkDistributionProps extends BaseSiteCdkDistributionProps {}
 export interface SsrSiteProps {
+  /**
+   * Bind resources for the function
+   *
+   * @example
+   * ```js
+   * new Function(stack, "Function", {
+   *   handler: "src/function.handler",
+   *   bind: [STRIPE_KEY, bucket],
+   * })
+   * ```
+   */
+  bind?: SSTConstruct[];
   /**
    * The SSR function is deployed to Lambda in a single region. Alternatively, you can enable this option to deploy to Lambda@Edge.
    * @default false
@@ -208,7 +220,7 @@ export class SsrSite extends Construct implements SSTConstruct {
      */
     certificate?: acm.ICertificate;
   };
-  protected props: Omit<SsrSiteProps, "path"> & { path: string; };
+  protected props: Omit<SsrSiteProps, "path"> & { path: string };
   /**
    * Determines if a placeholder site should be deployed instead. We will set
    * this to `true` by default when performing local development, although the
@@ -400,10 +412,9 @@ export class SsrSite extends Construct implements SSTConstruct {
       },
       permissions: {
         "ssm:GetParameters": [
-          `arn:${Stack.of(this).partition}:ssm:${app.region}:${app.account}:parameter${getParameterPath(
-            this,
-            "url"
-          )}`,
+          `arn:${Stack.of(this).partition}:ssm:${app.region}:${
+            app.account
+          }:parameter${getParameterPath(this, "url")}`,
         ],
       },
     };
@@ -497,9 +508,9 @@ export class SsrSite extends Construct implements SSTConstruct {
     const app = this.node.root as App;
     const script = path.resolve(__dirname, "../support/base-site-archiver.mjs");
     const fileSizeLimit = app.isRunningSSTTest()
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore: "sstTestFileSizeLimitOverride" not exposed in props
-      ? this.props.sstTestFileSizeLimitOverride || 200
+      ? // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-ignore: "sstTestFileSizeLimitOverride" not exposed in props
+        this.props.sstTestFileSizeLimitOverride || 200
       : 200;
     const result = spawn.sync(
       "node",
@@ -575,11 +586,13 @@ export class SsrSite extends Construct implements SSTConstruct {
   }
 
   private createS3AssetFileOptionsForStub() {
-    return [{
-      exclude: "*",
-      include: "*",
-      cacheControl: "public,max-age=0,s-maxage=31536000,must-revalidate",
-    }];
+    return [
+      {
+        exclude: "*",
+        include: "*",
+        cacheControl: "public,max-age=0,s-maxage=31536000,must-revalidate",
+      },
+    ];
   }
 
   private createS3Bucket(): s3.Bucket {
@@ -601,7 +614,10 @@ export class SsrSite extends Construct implements SSTConstruct {
     }
   }
 
-  private createS3Deployment(assets: s3Assets.Asset[], fileOptions: { exclude: string, include: string, cacheControl: string }[]): CustomResource {
+  private createS3Deployment(
+    assets: s3Assets.Asset[],
+    fileOptions: { exclude: string; include: string; cacheControl: string }[]
+  ): CustomResource {
     // Create a Lambda function that will be doing the uploading
     const uploader = new lambda.Function(this, "S3Uploader", {
       code: lambda.Code.fromAsset(
@@ -897,8 +913,8 @@ export class SsrSite extends Construct implements SSTConstruct {
     const waitForInvalidation = this.isPlaceholder
       ? false
       : this.props.waitForInvalidation === false
-        ? false
-        : true;
+      ? false
+      : true;
     return new CustomResource(this, "CloudFrontInvalidation", {
       serviceToken: invalidator.functionArn,
       resourceType: "Custom::SSTCloudFrontInvalidation",